add docstrings

+++ magnetic_levitation/src/linear_interp.h
@@ -1,5 +1,10 @@
-
+// NOTE: arrays must be in increasing order ////////////////////////////
 int linear_interp(int x_val, int* x_data, int* y_data, unsigned int n_data_points) {
+    // interpolate between two points for known sets of x and y data
+    // x_val: value to do y interpolation for
+    // x_data, y_data: arrays of known coordinates in x and y
+    // n_data_points: lengths of both of the arrays
+
     unsigned int idx = 0;
     for (; idx < n_data_points; ++idx) {
         // iterate until x_data is exceeded by one of the data points we have.
@@ -13,10 +18,11 @@ int linear_interp(int x_val, int* x_data, int* y_data, unsigned int n_data_point
     } else if (idx == n_data_points - 1) {  // return the highest value in the range
         return x_data[n_data_points - 1];
     } else {  // compute the linear interpolation
-        return float(x_val - x_data[idx - 1]) / float(x_data[idx] - x_data[idx - 1]) * float(y_data[idx] - y_data[idx - 1]) + y_data[idx - 1];
+        return int(((float(x_val - x_data[idx - 1]) / float(x_data[idx] - x_data[idx - 1])) * float(y_data[idx] - y_data[idx - 1])) + y_data[idx - 1]);
     }
 }
 
+// verbatim copy of above, just using floats
 float linear_interp(float x_val, float* x_data, float* y_data, unsigned int n_data_points) {
     unsigned int idx = 0;
     for (; idx < n_data_points; ++idx) {
@@ -31,6 +37,6 @@ float linear_interp(float x_val, float* x_data, float* y_data, unsigned int n_da
     } else if (idx == n_data_points - 1) {  // return the highest value in the range
         return x_data[n_data_points - 1];
     } else {  // compute the linear interpolation
-        return float(x_val - x_data[idx - 1]) / float(x_data[idx] - x_data[idx - 1]) * float(y_data[idx] - y_data[idx - 1]) + y_data[idx - 1];
+        return (((float(x_val - x_data[idx - 1]) / float(x_data[idx] - x_data[idx - 1])) * float(y_data[idx] - y_data[idx - 1])) + y_data[idx - 1]);
     }
 }
